+++ CS361-GISTool-Group23/imagecompar.cpp
